lib/x86/adler32: remove the AVX-512BW implementation
This code doesn't seem worthwhile to keep around, given that using zmm
registers incurs a frequency drop.  On some CPU models, it is a very
substantial drop.  So it might seem worthwhile in microbenchmarks, but
the benefit is lost by the effect on other workloads.  Also in Alder
Lake, Intel is removing AVX-512 from client CPUs.  There's also room for
improvement in the AVX-2 implementation.

This code could come back later, but let's drop it for now.