Fix libdeflate.pc on Windows
Use Cflags.private to make the pkgconfig file specify the appropriate
flag when linking to libdeflate.dll on Windows.  To do this cleanly,
also invert the meaning of the flag, replacing LIBDEFLATE_DLL with
LIBDEFLATE_STATIC.  So now the DLL is the default, and LIBDEFLATE_STATIC
has to be defined when linking to the static library on Windows.

Fixes https://github.com/ebiggers/libdeflate/issues/229