Use MinGW naming convention for Windows libraries
For the import and static libraries, use the MinGW naming convention
(libdeflate.dll.a and libdeflate.a, respectively) instead of the Visual
Studio naming convention (libdeflate.lib and libdeflatestatic.lib,
respectively).  The original assumption was that the native Windows
convention was better.  However, instead people expect the naming
convention to reflect the toolchain with which the files were built.

Also change 'make install' to install the DLL into the bin directory
instead of the lib directory, again following the MinGW convention.

Fixes https://github.com/ebiggers/libdeflate/issues/190