lib/x86: make PCLMUL crc32 code mostly work with MSVC
Rework this code to avoid gcc vector extensions, and once that is done
allow it to be built with MSVC (and clang in MSVC compatibility mode).

One part is still missing, which is that the PCLMUL+AVX version isn't
actually compiled with AVX code generation enabled, as that would
require that it be put into a separate translation unit.

(Inspired by changes by Shawn Hoffman <godisgovernment@gmail.com>.)