deflate_decompress: return error when too many codeword lengths
When there are more encoded litlen and offset codeword lengths than
there are supposed to be based on the earlier fields, return
LIBDEFLATE_BAD_DATA instead of ignoring the extra lengths.

This isn't very important, but this aligns the behavior with zlib.

Also add a regression test for this change.

Update https://github.com/ebiggers/libdeflate/issues/288