lib/cpu_features: fix intrinsic support detection with latest clang
Unfortunately, checks like __has_builtin(__builtin_ia32_pclmulqdq128)
have stopped working properly with clang, starting with clang 15 on
x86_64 and clang 16 on arm64.  They now only expand to '1' when the
corresponding CPU feature is enabled on the command line.  As a result,
a lot of optimized code (mainly for the checksums) is not used.

The reason that libdeflate is doing these checks in the first place is
just to exclude old compiler versions that don't support the
corresponding set of intrinsics.  So, let's do it the "bad" way and
instead check the clang version number directly...