lib/x86/cpu_features: ensure xgetbv not executed when unsupported
Mark the asm block containing the xgetbv instruction with 'volatile'
so that compiler optimizations can't move it out from under the check
for OSXSAVE support.

This isn't known to have actually been happening, but it's theoretically
possible.  (It did happen in another project where I use a modified
version of this code that was more friendly to compiler optimizations.)