Add support for per-(de)compressor malloc/free functions
Allow specifying a custom memory allocator per-(de)compressor, to
support use cases where there might be different users of libdeflate in
the same process who want to use different memory allocators.

This is a replacement for
https://github.com/ebiggers/libdeflate/pull/312.  I changed the design
slightly to be more extensible.  The allocator is specified by a new
'struct libdeflate_options' which is extensible, so the same struct will
also work for special compression options/hints, which is something that
has been requested in libdeflate before.  (I'm a bit hesitant to add
such options, but this at least makes a place to put them if it's ever
done.)  I also made the pointer to the free function be stored in the
(de)compressor struct, so the API for freeing is unchanged.