lib/x86: simplify checks for intrinsics
Since all the x86 HAVE_*_INTRIN values are now, in practice, determined
by compiler version checks, it's simpler to just fold these checks
directly into the code that actually uses the intrinsics (*_impl.h).
This allows many of the conditions to be merged together.

This approach does mean that the conditions for using a particular type
of intrinsic may get duplicated in different places as more functions
get added.  This is partly why I used the layer of indirection
originally.  But in practice this hasn't been much of a problem.  For
crc32_x86_vpclmulqdq_avx2() we also have to check the compiler version
anyway because MSVC supports both VPCLMULQDQ and AVX2, but not together.