lib/x86: adding missing MSVC and Apple clang version checks
Most of the recently added code that uses newer x86 intrinsics (AVX-512,
VPCLMULQDQ, AVX-VNNI) is being compiled unconditionally when the
compiler is MSVC or Apple clang.  This is inconsistent with the minimum
versions of these compilers that we claim to support.  Therefore, add
the needed checks against _MSC_VER and __apple_build_version__.