Drop support for very old versions of gcc and clang
Drop support for gcc before v4.9 and clang before v3.9.  This will allow
simplifying some code.  For example, we'll be able to assume that if the
compiler is gcc or clang, then on x86 the target function attribute will
work properly and intrinsics up to AVX2 will be supported.

Document and start explicitly enforcing the minimum compiler versions.

For MSVC, start enforcing a minimum of Visual Studio 2015.  However, I
believe this was already required due to the use of stdbool.h.