lib/x86: simplify by not trying to skip target attributes
It's true that if e.g. __AVX2__ is defined then it's unnecessary to use
__attribute__((target("avx2"))) on functions that use AVX-2 intrinsics.
But it seems to be harmless to add it.  All the target options we use on
x86 seem to work additively, so at worst it's a no-op.  We also no
longer support any gcc or clang versions that don't support the target
function attribute.  Therefore, simplify the code by not skipping using
the target function attribute when all the options are already enabled
in the main target.

Note that this change has no effect on MSVC builds, as MSVC doesn't
support the target function attribute anyway.