lib/x86/crc32: handle alignment differently
On recent x86 processors, vmovdqu with an aligned pointer is just as
fast as vmovdqa.  The recently updated x86 Adler-32 code takes advantage
of this by always using vmovdqu, and only aligning the pointer if the
length is large.  This commit updates the x86 CRC-32 code to use a
similar strategy.  This reduces the code length, since it's no longer
necessary to have a 4*VL-wide loop with vmovdqu.  Instead now there's
just special handling of len < 8*VL, and an 8*VL-wide loop with vmovdqu.