lib/x86: Specify evex512 and no-evex512 when appropriate
libdeflate failed to build with -mno-evex512 with clang 18+ or gcc 14+
because that flag causes the evex512 feature (which is new in those
compiler versions) to have to be explicitly added to functions that use
AVX512 with 512-bit vectors.  Add this feature where appropriate.

For the same compiler versions, also add no-evex512 to the functions
that use AVX512 + 256-bit vectors so that they are ready for AVX10/256.

Resolves https://github.com/ebiggers/libdeflate/issues/394