Add basic support for Windows ARM64EC
Make libdeflate build for ARM64EC, and add ARM64EC to GitHub Actions.

ARM64EC pretends to be x86_64, but lots of x86_64 functionality is
broken:

- 256-bit and 512-bit vectors are unsupported
- _mm*_load*() do not accept void pointers
- _m_prefetchw() is not supported
- bmi2 intrinsics are not supported

For now just do the simple thing and treat ARM64EC as a generic
architecture, so do not try to use any of the x86_64 optimized code.

The premise of optimizing for ARM64EC seems questionable, after all,
given that proper ARM64 builds are fully supported.